Changed random weight deltas to be gaussian with a sigma of 0.01 instead of 0.1. This matches SharpNEAT 2.x. The intention for the SharpNEAT 4.0 release was to keep parameters such as these identical to SharpNEAT 2.x, so that the core neuroevolution algorithm is as close as possible between the two versions, despite the big architectural overhaul.
This allows us to compare the performance/efficacy of 2.x and 4.x, to give some assurance that there isn't some bug/flaw in the 4.x code base that impacts performance.

Future releases can then be free to explore tuning of hyper parameters and such, once we have established that 4.x performs at least as well as 2.x.